Exiting the container falls back to localhost
`kubectl exec -it` starts an interactive session with the container. Exiting the container (`
root@helloworld-...:/app/src# exit`) will throw the user out of the session to the computer kubectl was executed on. Running `hostname -i` afterwards therefore does not show the node's IP. It shows the IP of the computer kubectl was run on.

Removed the lines to not cause confusion.